code climate suggested changes
