+++ app/both/utils.js
@@ -1,24 +1,24 @@
-const checkUserPermissions = function({
+export const checkUserPermissions = function({
   schemaType,
   action,
   userId,
   doc,
-  hasDeparted
+  hasDeparted,
 }) {
   let userIsAdministrator;
   let isHomeManagedByUser = false;
   if (!userId) return false;
 
   // Check if user is administrator
-  userIsAdministrator = Roles.userIsInRole(userId, ["admin"]);
+  userIsAdministrator = Roles.userIsInRole(userId, ['admin']);
   if (userIsAdministrator) return true;
 
   if (!doc.homeId) return false;
 
   // If user is not admin, check if he is manager
-  isHomeManagedByUser = Meteor.call("isHomeManagedByUser", {
+  isHomeManagedByUser = Meteor.call('isHomeManagedByUser', {
     userId,
-    homeId: doc.homeId
+    homeId: doc.homeId,
   });
 
   /* Check if current action is permitted.
@@ -28,15 +28,15 @@ const checkUserPermissions = function({
   let canEdit = isHomeManagedByUser;
 
   // Managers cannot delete
-  if (isHomeManagedByUser && action === "remove") {
+  if (isHomeManagedByUser && action === 'remove') {
     canEdit = false;
   }
 
   // Manager cannot update an inactive residency
   if (
     isHomeManagedByUser &&
-    schemaType === "residency" &&
-    action === "update" &&
+    schemaType === 'residency' &&
+    action === 'update' &&
     hasDeparted
   ) {
     canEdit = false;
@@ -44,6 +44,17 @@ const checkUserPermissions = function({
   return canEdit;
 };
 
-module.exports = {
-  checkUserPermissions
+export const getActiveResidency = function(residentId) {
+  return Residencies.findOne({
+    $and: [
+      {
+        residentId,
+      },
+      {
+        moveOut: {
+          $exists: false,
+        },
+      },
+    ],
+  });
 };
